chore: move accounts dao to api module
