+++ persistence-api/src/commonMain/kotlin/com/wire/kalium/persistence/model/AuthSessionEntity.kt
