+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/dao/UserDAO.kt
@@ -18,35 +18,11 @@
 
 package com.wire.kalium.persistence.dao
 
-import com.wire.kalium.logger.obfuscateDomain
-import com.wire.kalium.logger.obfuscateId
 import com.wire.kalium.persistence.dao.ManagedByEntity.WIRE
 import kotlinx.coroutines.flow.Flow
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-
-@Serializable
-data class QualifiedIDEntity(
-    @SerialName("value") val value: String,
-    @SerialName("domain") val domain: String
-) {
-    override fun toString(): String = if (domain.isEmpty()) value else "$value${VALUE_DOMAIN_SEPARATOR}$domain"
-
-    fun toLogString(): String = if (domain.isEmpty()) {
-        value.obfuscateId()
-    } else {
-        "${value.obfuscateId()}${VALUE_DOMAIN_SEPARATOR}${domain.obfuscateDomain()}"
-    }
-
-    companion object {
-        private const val VALUE_DOMAIN_SEPARATOR = '@'
-    }
-
-}
-
-typealias UserIDEntity = QualifiedIDEntity
-typealias ConversationIDEntity = QualifiedIDEntity
 
 @Serializable
 enum class SupportedProtocolEntity {
@@ -190,14 +166,6 @@ enum class UserTypeEntity {
      * when current user doesn't belongs to any team
      */
     NONE;
-}
-
-/**
- * This is used to indicate if the self user (account) is managed by SCIM or Wire
- * If the user is managed by other than [WIRE], then is a read only account.
- */
-enum class ManagedByEntity {
-    WIRE, SCIM
 }
 
 internal typealias UserAssetIdEntity = QualifiedIDEntity
