+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/daokaliumdb/AccountsDAO.kt
@@ -22,7 +22,6 @@ import app.cash.sqldelight.coroutines.asFlow
 import com.wire.kalium.persistence.AccountsQueries
 import com.wire.kalium.persistence.CurrentAccountQueries
 import com.wire.kalium.persistence.dao.ManagedByEntity
-import com.wire.kalium.persistence.dao.QualifiedIDEntity
 import com.wire.kalium.persistence.dao.UserIDEntity
 import com.wire.kalium.persistence.model.LogoutReason
 import com.wire.kalium.persistence.model.ServerConfigEntity
@@ -33,151 +32,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.CoroutineContext
-
-data class AccountInfoEntity(
-    val userIDEntity: UserIDEntity,
-    val logoutReason: LogoutReason?
-)
-
-data class PersistentWebSocketStatusEntity(
-    val userIDEntity: UserIDEntity,
-    val isPersistentWebSocketEnabled: Boolean
-)
-
-data class FullAccountEntity(
-    val info: AccountInfoEntity,
-    val serverConfigId: String,
-    val ssoId: SsoIdEntity?,
-    val persistentWebSocketStatusEntity: PersistentWebSocketStatusEntity,
-    val managedBy: ManagedByEntity?
-)
-
-@Suppress("FunctionParameterNaming", "LongParameterList")
-internal object AccountMapper {
-    fun fromAccount(
-        user_id: UserIDEntity,
-        logout_reason: LogoutReason?,
-    ): AccountInfoEntity = AccountInfoEntity(
-        userIDEntity = user_id,
-        logoutReason = logout_reason,
-    )
-
-    fun fromPersistentWebSocketStatus(
-        user_id: UserIDEntity,
-        isPersistentWebSocketEnabled: Boolean,
-    ): PersistentWebSocketStatusEntity = PersistentWebSocketStatusEntity(
-        userIDEntity = user_id,
-        isPersistentWebSocketEnabled = isPersistentWebSocketEnabled
-    )
-
-    fun fromFullAccountInfo(
-        id: QualifiedIDEntity,
-        scim_external_id: String?,
-        subject: String?,
-        tenant: String?,
-        server_config_id: String,
-        logout_reason: LogoutReason?,
-        isPersistentWebSocketEnabled: Boolean,
-        managedBy: ManagedByEntity?
-    ): FullAccountEntity = FullAccountEntity(
-        info = fromAccount(id, logout_reason),
-        serverConfigId = server_config_id,
-        ssoId = toSsoIdEntity(scim_external_id, subject, tenant),
-        persistentWebSocketStatusEntity = fromPersistentWebSocketStatus(id, isPersistentWebSocketEnabled),
-        managedBy = managedBy
-    )
-
-    fun toSsoIdEntity(
-        scim_external_id: String?,
-        subject: String?,
-        tenant: String?
-    ): SsoIdEntity? = if (scim_external_id == null && subject == null && tenant == null) {
-        null
-    } else {
-        SsoIdEntity(scim_external_id, subject, tenant)
-    }
-
-    fun fromUserIDWithServerConfig(
-        userId: QualifiedIDEntity,
-        serverConfigId: String,
-        apiBaseUrl: String,
-        accountBaseUrl: String,
-        webSocketBaseUrl: String,
-        blackListUrl: String,
-        teamsUrl: String,
-        websiteUrl: String,
-        isOnPremises: Boolean,
-        domain: String?,
-        commonApiVersion: Int,
-        federation: Boolean,
-        apiProxyHost: String?,
-        apiProxyPort: Int?,
-        apiProxyNeedsAuthentication: Boolean?,
-        title: String
-    ): Pair<UserIDEntity, ServerConfigEntity> {
-        val apiProxy: ServerConfigEntity.ApiProxy? =
-            if (apiProxyHost != null && apiProxyPort != null && apiProxyNeedsAuthentication != null) {
-                ServerConfigEntity.ApiProxy(apiProxyNeedsAuthentication, apiProxyHost, apiProxyPort)
-            } else {
-                null
-            }
-
-        val serverConfig = ServerConfigEntity(
-            id = serverConfigId,
-            links = ServerConfigEntity.Links(
-                api = apiBaseUrl,
-                accounts = accountBaseUrl,
-                webSocket = webSocketBaseUrl,
-                blackList = blackListUrl,
-                teams = teamsUrl,
-                website = websiteUrl,
-                isOnPremises = isOnPremises,
-                apiProxy = apiProxy,
-                title = title
-            ),
-            metaData = ServerConfigEntity.MetaData(
-                domain = domain,
-                apiVersion = commonApiVersion,
-                federation = federation,
-            )
-        )
-
-        return userId to serverConfig
-    }
-
-}
-
-@Suppress("TooManyFunctions")
-interface AccountsDAO {
-    suspend fun ssoId(userIDEntity: UserIDEntity): SsoIdEntity?
-    suspend fun insertOrReplace(
-        userIDEntity: UserIDEntity,
-        ssoIdEntity: SsoIdEntity?,
-        serverConfigId: String,
-        isPersistentWebSocketEnabled: Boolean
-    )
-
-    suspend fun observeAccount(userIDEntity: UserIDEntity): Flow<AccountInfoEntity?>
-    suspend fun allAccountList(): List<AccountInfoEntity>
-    suspend fun allValidAccountList(): List<AccountInfoEntity>
-    fun observerValidAccountList(): Flow<List<AccountInfoEntity>>
-    suspend fun observeAllAccountList(): Flow<List<AccountInfoEntity>>
-    suspend fun isFederated(userIDEntity: UserIDEntity): Boolean?
-    suspend fun doesValidAccountExists(userIDEntity: UserIDEntity): Boolean
-    suspend fun currentAccount(): AccountInfoEntity?
-    fun observerCurrentAccount(): Flow<AccountInfoEntity?>
-    suspend fun setCurrentAccount(userIDEntity: UserIDEntity?)
-    suspend fun updateSsoIdAndScimInfo(userIDEntity: UserIDEntity, ssoIdEntity: SsoIdEntity?, managedBy: ManagedByEntity?)
-    suspend fun deleteAccount(userIDEntity: UserIDEntity)
-    suspend fun markAccountAsInvalid(userIDEntity: UserIDEntity, logoutReason: LogoutReason)
-    suspend fun updatePersistentWebSocketStatus(userIDEntity: UserIDEntity, isPersistentWebSocketEnabled: Boolean)
-    suspend fun persistentWebSocketStatus(userIDEntity: UserIDEntity): Boolean
-    suspend fun accountInfo(userIDEntity: UserIDEntity): AccountInfoEntity?
-    fun fullAccountInfo(userIDEntity: UserIDEntity): FullAccountEntity?
-    suspend fun getAllValidAccountPersistentWebSocketStatus(): Flow<List<PersistentWebSocketStatusEntity>>
-    suspend fun getAccountManagedBy(userIDEntity: UserIDEntity): ManagedByEntity?
-    suspend fun validAccountWithServerConfigId(): Map<UserIDEntity, ServerConfigEntity>
-}
 
 @Suppress("TooManyFunctions")
 internal class AccountsDAOImpl internal constructor(
